stats.h: Removed unneeded bodge. Fixed two copiler warnings.
git-svn-id: file:///home/xapian-svn/xapian/branches/branch-newmatcher@2434 6b2f1b92-a4f3-0310-943f-de3cbe41748b

+++ xapian-core/matcher/mergepostlist.h
@@ -121,7 +121,7 @@ inline bool
 MergePostList::at_end() const
 {
     Assert(current != -1);
-    return current >= plists.size();
+    return (unsigned int)current >= plists.size();
 }
 
 inline std::string
